+++ alice-bootstrap/src/main/java/module-info.java
@@ -1,9 +1,6 @@
 module alice.agent.app.main {
   exports org.cland.alice.agent;
 
-  requires alice.agent.alice.core.agent.main;
-  requires alice.agent.alice.model.main;
-  requires alice.agent.command.main;
   requires alice.agent.facade.cmd.main;
   requires alice.agent.facade.tui.main;
   requires org.slf4j;